[FIX](Function Architecture): Fixed function architecture and divided function (helper function in SDLUtils and main function in SDL.cpp)
